requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
faster-whisper>=1.1.0
google-genai>=1.0.0
langgraph>=0.1.14
langchain>=0.2.6
langchain-core>=0.2.10
requests>=2.32.3
python-dotenv>=1.0.1
pydantic>=2.7.4
